Move Allele to TextFlow utility methods to new class
